extract private classes to external inner classes
