+++ butterflysdk/src/main/java/com/butterfly/sdk/Logic/WebViewerActivity.kt
@@ -1,4 +1,4 @@
-package com.butterfly.sdk
+package com.butterfly.sdk.Logic
 
 import android.annotation.SuppressLint
 import android.app.Activity
@@ -7,20 +7,11 @@ import android.content.pm.PackageManager
 import android.content.res.Resources
 import android.graphics.Color
 import android.net.Uri
-import android.os.Build
 import android.os.Bundle
-import android.os.Handler
-import android.os.HandlerThread
-import android.os.Looper
 import android.view.Gravity
 import android.view.View
 import android.view.ViewGroup
-import android.webkit.JavascriptInterface
-import android.webkit.WebResourceError
-import android.webkit.WebResourceRequest
-import android.webkit.WebResourceResponse
 import android.webkit.WebView
-import android.webkit.WebViewClient
 import android.widget.Button
 import android.widget.RelativeLayout
 import android.widget.TextView
@@ -28,16 +19,13 @@ import com.butterfly.sdk.utils.EventBus
 import com.butterfly.sdk.utils.SdkLogger
 import com.butterfly.sdk.utils.Utils
 import org.json.JSONObject
-import java.io.IOException
-import java.io.OutputStreamWriter
-import java.net.HttpURLConnection
+import java.net.MalformedURLException
 import java.net.URL
 import java.util.Locale
-import javax.net.ssl.HttpsURLConnection
 
 class WebViewerActivity: Activity(), EventBus.Listener {
     class AbortEvent(val caller: Activity) : EventBus.Event()
-    private object IntentExtraKeys {
+    internal object IntentExtraKeys {
         const val URL = "url"
         const val SHOULD_CLEAR_CACHE = "shouldClearCache"
     }
@@ -101,6 +89,22 @@ class WebViewerActivity: Activity(), EventBus.Listener {
             )
         }
 
+        fun open(activity: Activity, uri: Uri, apiKey: String) {
+            Utils.saveContext(activity)
+            if (apiKey.isEmpty()) return
+
+            try {
+                val url = URL(uri.scheme, uri.host, uri.path)
+                // extract url params
+                // throttling
+                // fetchButterflyParamsFromURL
+                // extractURLExtraParamsFromDictionary
+                // openReporterUsingKey
+            } catch (e: MalformedURLException) {
+                return
+            }
+        }
+
         val TAG: String get() {
             return WebViewerActivity::class.java.simpleName
         }
@@ -108,7 +112,7 @@ class WebViewerActivity: Activity(), EventBus.Listener {
         var countryCodeToOverride: String? = null
         var customColorHexaString: String? = null // May be: "0xFF91BA48", "FF91BA48", "91BA48"
 
-        private val urlWhiteList: HashSet<String> = HashSet()
+        internal val urlWhiteList: HashSet<String> = HashSet()
     }
 
     private var token: EventBus.ListenerToken? = null
@@ -309,58 +313,6 @@ class WebViewerActivity: Activity(), EventBus.Listener {
         super.onDestroy()
     }
 
-    private class ButterflyWebViewClient(val navigationRequestsListener: NavigationRequestsListener) : WebViewClient() {
-        override fun onLoadResource(view: WebView?, url: String?) {
-            super.onLoadResource(view, url)
-        }
-
-        override fun onReceivedError(
-            view: WebView?,
-            request: WebResourceRequest?,
-            error: WebResourceError?
-        ) {
-            super.onReceivedError(view, request, error)
-//            handler("page error")
-        }
-
-        override fun onReceivedHttpError(
-            view: WebView?,
-            request: WebResourceRequest?,
-            errorResponse: WebResourceResponse?
-        ) {
-            super.onReceivedHttpError(view, request, errorResponse)
-//            handler("page error")
-        }
-
-        override fun shouldInterceptRequest(view: WebView?, request: WebResourceRequest?): WebResourceResponse? {
-            return super.shouldInterceptRequest(view, request)
-        }
-
-        override fun shouldOverrideUrlLoading(view: WebView?, request: WebResourceRequest?): Boolean {
-            val ignoreUrl = true
-            val urlString = if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.LOLLIPOP) {
-                request?.url?.toString() ?: ""
-            } else {
-                request?.toString() ?: ""
-            }
-
-            if (urlString.isEmpty()) return ignoreUrl
-
-            navigationRequestsListener.onNavigationRequest(urlString)
-
-            if (isWhiteListed(urlString)) {
-                view?.loadUrl(urlString)
-                return ignoreUrl
-            }
-
-            return ignoreUrl
-        }
-
-        private fun isWhiteListed(urlString: String): Boolean {
-            return urlWhiteList.any { item -> urlString.startsWith(item) }
-        }
-    }
-
     @Deprecated("Deprecated in Java")
     override fun onBackPressed() {
         val initialUrl = this.initialUrl ?: ""
@@ -373,172 +325,6 @@ class WebViewerActivity: Activity(), EventBus.Listener {
 
     private fun Int.dpToPx(): Int {
         return (this * Resources.getSystem().displayMetrics.density).toInt()
-    }
-
-    private class AndroidJavascriptInterface(private val markAsHandled: (resultString: String, commandId: String) -> Unit) {
-        lateinit var host: WebViewerActivity
-
-        @JavascriptInterface
-        fun postMessage(messageFromJs: String) {
-            val messageJson = JSONObject(messageFromJs)
-            val commandName = messageJson.remove("commandName")?.toString() ?: ""
-            val commandId = messageJson.remove("commandId")?.toString() ?: ""
-
-            when (commandName) {
-                "sendRequest" -> {
-                    messageJson.remove("urlString")?.toString()?.let { urlString ->
-                        messageJson.remove("key")?.toString()?.let { apiKey ->
-                            val butterflyApiKey: String = if (Utils.isDebuggable() && !apiKey.startsWith("debug-")) {
-                                "debug-$apiKey"
-                            } else {
-                                apiKey
-                            }
-
-                            Communicator(urlString, messageJson, mapOf("butterfly_host_api_key" to butterflyApiKey)).call { networkResult ->
-                                var resultString = "error"
-                                if (networkResult == "OK") {
-                                    resultString = networkResult
-                                }
-                                markAsHandled.invoke(resultString, commandId)
-                            }
-                        }
-                    }
-                }
-
-                "navigateTo" -> {
-                    messageJson.remove("urlString")?.toString()?.let { urlString ->
-                        host.startActivity(
-                            Intent(host, WebViewerActivity::class.java).putExtra(
-                                IntentExtraKeys.URL,
-                                urlString
-                            )
-                        )
-                        markAsHandled("OK", commandId)
-                    } ?: run {
-                        markAsHandled("", commandId)
-                    }
-                }
-
-                "allowNavigation" -> {
-                    messageJson.remove("urlString")?.toString()?.let { urlString ->
-                        urlWhiteList.add(urlString)
-                        markAsHandled.invoke("OK", commandId)
-                    } ?: run {
-                        markAsHandled("", commandId)
-                    }
-                }
-
-                "open" -> {
-                    try {
-                        val url = messageJson.getJSONArray("components").get(0).toString()
-                        if (url.isNotEmpty()) {
-                            val intent = Intent(Intent.ACTION_VIEW)
-                            intent.data = Uri.parse(url)
-                            host.startActivity(intent)
-                            markAsHandled.invoke("OK", commandId)
-                        } else {
-                            markAsHandled("", commandId)
-                        }
-                    } catch (e: Throwable) {
-                        SdkLogger.error(TAG, e)
-                        markAsHandled("", commandId)
-                    }
-                }
-
-                else -> {
-                    markAsHandled("", commandId)
-                }
-            }
-        }
-    }
-
-    private class Communicator(private val urlString: String, private val requestBody: JSONObject? = null, private val headers: Map<String, String> = mapOf()) {
-        companion object {
-
-            private val bgThreadHandler: Handler by lazy {
-                val handlerThread = HandlerThread("BFCommunicator Thread")
-                handlerThread.start()
-                val handler = Handler(handlerThread.looper)
-
-                handler
-            }
-
-            // Ref: https://stackoverflow.com/a/3584332/2735029
-            fun pingUrl(url: String, timeout: Int, callback: (Boolean) -> Unit) {
-//                var url = url
-//                url = url.replaceFirst(
-//                    "^https".toRegex(),
-//                    "http"
-//                ) // (???) Otherwise an exception may be thrown on invalid SSL certificates.
-                val looper = Looper.myLooper() ?: return
-                val callerHandler = Handler(looper)
-                bgThreadHandler.post {
-                    val isReachable = try {
-                        val connection: HttpURLConnection =
-                            URL(url).openConnection() as HttpURLConnection
-                        connection.connectTimeout = timeout
-                        connection.readTimeout = timeout
-                        connection.requestMethod = "HEAD"
-                        val responseCode: Int = connection.responseCode
-                        responseCode in 200..399
-                    } catch (exception: IOException) {
-                        SdkLogger.error(TAG, exception)
-                        false
-                    }
-
-                    callerHandler.post {
-                        callback.invoke(isReachable)
-                    }
-                }
-            }
-        }
-
-        fun call(callback: (String) -> Unit) {
-            if(urlString.isEmpty()) {
-                callback.invoke("")
-                return
-            }
-
-            val url = URL(urlString)
-            val looper = Looper.myLooper() ?: return
-            val callerHandler = Handler(looper)
-            bgThreadHandler.post {
-                try {
-                    var responseString = ""
-                    // Thanks a lot to: https://johncodeos.com/post-get-put-delete-requests-with-httpurlconnection/
-                    val connection = url.openConnection() as HttpsURLConnection
-                    connection.requestMethod = "POST"
-                    connection.setRequestProperty("Content-Type", "application/json") // The format of the content we're sending to the server
-                    headers.forEach { headerEntry ->
-                        connection.setRequestProperty(headerEntry.key, headerEntry.value)
-                    }
-                    connection.setRequestProperty("Accept", "application/json") // The format of response we want to get from the server
-                    connection.doInput = true
-                    connection.doOutput = true
-
-                    // Send the JSON we created
-                    val outputStreamWriter = OutputStreamWriter(connection.outputStream)
-                    outputStreamWriter.write(requestBody.toString())
-                    outputStreamWriter.flush()
-
-                    // Check if the connection is successful
-                    val responseCode = connection.responseCode
-                    if (responseCode == HttpsURLConnection.HTTP_OK) {
-                        responseString = connection.inputStream.bufferedReader()
-                            .use { it.readText() }  // defaults to UTF-8
-//                        Log.d("Pretty Printed JSON :", responseString)
-                    } else {
-                        //Log.e("HTTPS URL CONNECTION ERROR", responseCode.toString())
-                    }
-
-                    callerHandler.post {
-                        callback.invoke(responseString)
-                    }
-                } catch (e: Throwable) {
-//                    e.printStackTrace()
-                }
-            }
-        }
     }
 
     override fun onEvent(event: EventBus.Event) {
